fix copper refilling
currently the copper refilling does not takes into account the
mouse bites, thus there is no clearance around the holes.
this is because their zone is not added in copperFillNonBoardAreas.
to be sure such issues don't occur, now all zones are refilled.
not sure if https://gitlab.com/kicad/code/kicad/-/issues/11666 has
been fixed, but I was not able to reproduce it even after 200 runs.
in case it is a in-memory issue, the board is still saved and
reloaded before refill.