Add a few user layers, e.g. to choose as V-Cut layer
If the user wants to reserve a dedicated (new) layer for V-Cut lines
(which is always a good idea), it would be nice to be able to choose one
of the User layers for this purpose, as the other layers may already be
in use by the source layout.
More User layers could be added in the future.