OpenPNP ignore footprints not in pos file
The `naturalComponentKey()` method will fail on non-annotated footprints (i.e. `REF***`). This can happens when adding a graphical footprint in the PCB without it being in the schematic (which will commonly be marked as `Not In schematic` and `Exclude from position file`).

This commit ignores footprints that are excluded from the position file as they will have no bearing on the OpenPNP pos file process, and they might cause the process to fail for the reasons explained above.