Update windows.md with troubleshooting instructions
The batch script that initializes paths in the command prompt enumerates existing directories only. Some plugin and script paths are only created after pip installs files that need to belong there. Thus, if KiKit is the first such plugin the user installs, the "test that it works" section will fail only because the CLI needs to be restarted to pick up the new paths.